Conflicts:
	trunk/ServerCore/Stump.Server.WorldServer/Handlers/Approach/ApproachHandler.cs
	trunk/ServerCore/Stump.Server.WorldServer/Handlers/Character/CharacterChoiceHandler.cs
	trunk/ServerCore/Stump.Server.WorldServer/Handlers/Character/CharacterCreationHandler.cs

+++ trunk/ServerCore/Stump.Server.WorldServer/Data/ItemLoader.cs
@@ -15,7 +15,7 @@ namespace Stump.Server.WorldServer.Data
             foreach (var item in itemTemplates)
             {
                 itemTemplates.Where(entry => entry.Value.Id == item.Value.Id)
-                    .First().Value.Name = DataLoader.I18NFile.ReadText(item.Value.NameId);
+                    .First().Value.Name = DataLoader.I18NFile.ReadText((int) item.Value.NameId);
             }
         }
