+++ trunk/ServerCore/Stump.Server.WorldServer/Items/Inventory.cs
@@ -329,27 +329,27 @@ namespace Stump.Server.WorldServer.Items
         private void OnItemMove(Inventory sender, Item item, CharacterInventoryPositionEnum oldPosition)
         {
             // Update entity skin
-            List<short> newskins = Owner.Skins.Take(1).ToList();
+            List<uint> newskins = Owner.Skins.Take(1).ToList();
 
             Item hatItem;
             if ((hatItem = GetItem(CharacterInventoryPositionEnum.ACCESSORY_POSITION_HAT)) != null)
-                newskins.Add((short) hatItem.Template.AppearanceId);
+                newskins.Add((uint)hatItem.Template.AppearanceId);
 
             Item capeItem;
             if ((capeItem = GetItem(CharacterInventoryPositionEnum.ACCESSORY_POSITION_CAPE)) != null)
-                newskins.Add((short) capeItem.Template.AppearanceId);
+                newskins.Add((uint)capeItem.Template.AppearanceId);
 
             Item weaponItem;
             if ((weaponItem = GetItem(CharacterInventoryPositionEnum.ACCESSORY_POSITION_WEAPON)) != null)
-                newskins.Add((short) weaponItem.Template.AppearanceId);
+                newskins.Add((uint)weaponItem.Template.AppearanceId);
 
             Item petsItem;
             if ((petsItem = GetItem(CharacterInventoryPositionEnum.ACCESSORY_POSITION_PETS)) != null)
-                newskins.Add((short) petsItem.Template.AppearanceId);
+                newskins.Add((uint)petsItem.Template.AppearanceId);
 
             Item shieldItem;
             if ((shieldItem = GetItem(CharacterInventoryPositionEnum.ACCESSORY_POSITION_SHIELD)) != null)
-                newskins.Add((short) shieldItem.Template.AppearanceId);
+                newskins.Add((uint)shieldItem.Template.AppearanceId);
 
             Owner.Skins = newskins;
